test both kinds of external process
